+++ src/pages/userInfoInputPage/components/KeywordInput/KeywordInput.module.css
@@ -1,0 +1,27 @@
+.keyword {
+  display: inline-block;
+  width: 100%;
+  height: 40px;
+  padding: 9px 12px;
+
+  background-color: var(--white, #fff);
+  border: 1px solid var(--secondary, #1771b2);
+  border-radius: 37px;
+
+  color: var(--secondary, #1771b2);
+  font-size: var(--text-xs, 14px);
+  font-weight: 700;
+  line-height: normal;
+}
+
+input[type='checkbox']:checked + .keyword {
+  background-color: var(--secondary, #1771b2);
+  color: #b9d4e8;
+}
+
+.appearance-none {
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+  z-index: 1;
+}
